backend/requirements.txt
flask
flask-cors
cohere
transformers
requests
beautifulsoup4
scikit-learn
numpy
tqdm
python-dotenv
